+++ relay-metrics/src/statsd.rs
@@ -99,20 +99,18 @@ pub enum MetricHistograms {
     BucketsDelay,
 
     /// The number of batches emitted per partition by [`crate::aggregation::Aggregator`].
-    ///
-    /// This metric is only emitted if a partition key is set.
-    ///
-    /// Tags:
-    ///   - `partition_key`: The logical sharding key for the current batch.
     BatchesPerPartition,
 
     /// The number of buckets in a batch emitted by [`crate::aggregation::Aggregator`].
     ///
     /// This corresponds to the number of buckets that will end up in an envelope.
-    ///
-    /// Tags:
-    ///   - `partition_key`: The logical sharding key for the current batch.
     BucketsPerBatch,
+
+    /// Distribution of flush buckets over partition keys.
+    ///
+    /// The distribution of buckets should be even.
+    /// If it is not, this metric should expose it.
+    PartitionKeys,
 }
 
 impl HistogramMetric for MetricHistograms {
@@ -123,6 +121,7 @@ impl HistogramMetric for MetricHistograms {
             Self::BucketsDelay => "metrics.buckets.delay",
             Self::BatchesPerPartition => "metrics.buckets.batches_per_partition",
             Self::BucketsPerBatch => "metrics.buckets.per_batch",
+            Self::PartitionKeys => "metrics.buckets.partition_keys",
         }
     }
 }