+++ relay-profiling/src/android.rs
@@ -1,5 +1,7 @@
+use std::collections::{HashMap, HashSet};
+
 use android_trace_log::chrono::{DateTime, Utc};
-use android_trace_log::{AndroidTraceLog, Clock, Time};
+use android_trace_log::{AndroidTraceLog, Clock, Time, Vm};
 use serde::{Deserialize, Serialize};
 
 use relay_general::protocol::EventId;
@@ -53,18 +55,35 @@ struct AndroidProfile {
     #[serde(default, skip_serializing)]
     sampled_profile: String,
 
-    #[serde(default)]
-    profile: Option<AndroidTraceLog>,
+    #[serde(default = "AndroidProfile::default")]
+    profile: AndroidTraceLog,
 }
 
 impl AndroidProfile {
+    fn default() -> AndroidTraceLog {
+        AndroidTraceLog {
+            data_file_overflow: Default::default(),
+            clock: Clock::Global,
+            elapsed_time: Default::default(),
+            total_method_calls: Default::default(),
+            clock_call_overhead: Default::default(),
+            vm: Vm::Dalvik,
+            start_time: Utc::now(),
+            pid: Default::default(),
+            gc_trace: Default::default(),
+            threads: Default::default(),
+            methods: Default::default(),
+            events: Default::default(),
+        }
+    }
+
     fn parse(&mut self) -> Result<(), ProfileError> {
         let profile_bytes = match base64::decode(&self.sampled_profile) {
             Ok(profile) => profile,
             Err(_) => return Err(ProfileError::InvalidBase64Value),
         };
         self.profile = match android_trace_log::parse(&profile_bytes) {
-            Ok(profile) => Some(profile),
+            Ok(profile) => profile,
             Err(_) => return Err(ProfileError::InvalidSampledProfile),
         };
         Ok(())
@@ -79,6 +98,26 @@ impl AndroidProfile {
 
     fn has_transaction_metadata(&self) -> bool {
         !self.transaction_name.is_empty() && self.duration_ns > 0
+    }
+
+    /// Removes an event with a duration of 0
+    fn remove_events_with_no_duration(&mut self) {
+        let android_trace = &mut self.profile;
+        let events = &mut android_trace.events;
+        let clock = android_trace.clock;
+        let start_time = android_trace.start_time;
+        let mut timestamps_per_thread_id: HashMap<u16, HashSet<u64>> = HashMap::new();
+
+        for event in events.iter() {
+            let event_time = get_timestamp(clock, start_time, event.time);
+            timestamps_per_thread_id
+                .entry(event.thread_id)
+                .or_default()
+                .insert(event_time);
+        }
+
+        timestamps_per_thread_id.retain(|_, timestamps| timestamps.len() > 1);
+        events.retain(|event| timestamps_per_thread_id.contains_key(&event.thread_id));
     }
 }
 
@@ -103,7 +142,7 @@ pub fn expand_android_profile(payload: &[u8]) -> Result<Vec<Vec<u8>>, ProfileErr
         new_profile.set_transaction(transaction.clone());
         new_profile.transactions.clear();
 
-        let mut android_profile = new_profile.profile.as_ref().unwrap().clone();
+        let android_profile = new_profile.profile.clone();
         let clock = android_profile.clock;
         let start_time = android_profile.start_time;
         let mut events = android_profile.events;
@@ -113,9 +152,6 @@ pub fn expand_android_profile(payload: &[u8]) -> Result<Vec<Vec<u8>>, ProfileErr
             event_timestamp >= transaction.relative_start_ns
                 && event_timestamp <= transaction.relative_end_ns
         });
-
-        android_profile.events = events;
-        new_profile.profile = Some(android_profile);
 
         match serde_json::to_vec(&new_profile) {
             Ok(payload) => items.push(payload),
@@ -162,10 +198,6 @@ fn parse_android_profile(payload: &[u8]) -> Result<AndroidProfile, ProfileError>
     let mut profile: AndroidProfile =
         serde_json::from_slice(payload).map_err(ProfileError::InvalidJson)?;
 
-    if profile.sampled_profile.is_empty() {
-        return Ok(profile);
-    }
-
     if profile.transactions.is_empty() && !profile.has_transaction_metadata() {
         return Err(ProfileError::NoTransactionAssociated);
     }
@@ -176,9 +208,12 @@ fn parse_android_profile(payload: &[u8]) -> Result<AndroidProfile, ProfileError>
         }
     }
 
-    profile.parse()?;
+    if !profile.sampled_profile.is_empty() {
+        profile.parse()?;
+        profile.remove_events_with_no_duration();
+    }
 
-    if profile.profile.as_ref().unwrap().events.len() < 2 {
+    if profile.profile.events.is_empty() {
         return Err(ProfileError::NotEnoughSamples);
     }
 
@@ -209,6 +244,14 @@ mod tests {
     #[test]
     fn test_no_transaction() {
         let payload = include_bytes!("../tests/fixtures/profiles/android/no_transaction.json");
+        let data = parse_android_profile(payload);
+        assert!(data.is_err());
+    }
+
+    #[test]
+    fn test_remove_invalid_events() {
+        let payload =
+            include_bytes!("../tests/fixtures/profiles/android/remove_invalid_events.json");
         let data = parse_android_profile(payload);
         assert!(data.is_err());
     }